requirements.txt
aniso8601==8.0.0
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==3.0.6
django-cacheops==4.2
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-filter==2.2.0
django-graphql-jwt==0.3.1
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
funcy==1.14
graphene==2.1.8
graphene-django==2.10.0
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.3
Pillow==7.1.2
promise==2.3
psycopg2==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Rx==3.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
